+++ NEWS.md
@@ -1,7 +1,15 @@
-# bvars 1.0
+# bvars 1.1
 
 The package is under intensive development, and more functionality will be provided soon! To see the package [ROADMAP](https://github.com/bsvars/bvars/milestones) towards the next version.
 
 Have a question, or suggestion, or wanna get in touch? Email us at [contact\@bsvars.org](mailto:contact@bsvars.org) forum.
 
-* Initial CRAN submission.
+1. Updated package description with the paper by [Shang, Wang, Woźniak (2026)](https://doi.org/10.48550/arXiv.2608.28087)
+2. Introduced a range of code improvements by [Fei Shang](https://github.com/lcq110) in PRs [#15](https://github.com/bsvars/bvars/pull/15)-[#21](https://github.com/bsvars/bvars/pull/21) 
+3. Corrected documentation for argument `stationary` [#14](https://github.com/bsvars/bvars/issues/14)
+4. Introduced **R6** coding for class `Forecasts` [#11](https://github.com/bsvars/bvars/issues/11)
+5. Coded tests for package functions [#11](https://github.com/bsvars/bvars/issues/11)
+
+# bvars 1.0
+
+1. Initial CRAN submission.